# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 南加生活网 API
  version: 0.1.0
  description: |
    面向网站、管理后台及未来移动端的 REST API 合同。
    所有写操作必须支持幂等性、审计和统一 Problem Details 错误格式。
servers:
  - url: http://localhost:4000/v1
    description: Local
  - url: https://api.staging.example.com/v1
    description: Staging placeholder
security:
  - sessionCookie: []
...
paths:
...
  /auth/otp/request:
    post:
      tags: [Auth]
      security: []
      summary: Request an OTP challenge
      description: Always returns the same accepted projection for valid input; account existence is never disclosed.
      operationId: requestOtp
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "202":
          description: OTP accepted for delivery; response never reveals account existence
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpAcceptedResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /auth/otp/verify:
    post:
      tags: [Auth]
      security: []
      summary: Verify an OTP challenge
      description: Invalid, expired, consumed, and unknown challenges share one generic response.
      operationId: verifyOtp
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyRequest"
      responses:
        "200":
          description: Authenticated session established
          headers:
            Set-Cookie:
              schema:
                type: string
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/session:
    get:
      tags: [Auth]
      summary: Get the current session
      description: Resolves the opaque host-only session cookie; never returns the bearer token or contact identifiers.
      operationId: getSession
      responses:
        "200":
          description: Current session
          headers:
            Cache-Control:
              description: Session responses are never stored by shared or browser caches.
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [Auth]
      summary: Revoke the current session
      description: Idempotently revokes the presented session and expires the host-only cookie.
      operationId: logout
      responses:
        "204":
          description: Session revoked
          headers:
            Set-Cookie:
              description: Clears the Secure, HttpOnly, SameSite=Lax cookie scoped to /v1.
              schema: { type: string }
            Cache-Control:
              schema: { type: string, const: no-store }
...
  /admin/session:
    get:
      tags: [Admin]
      summary: Resolve the current operator console session
      description: |
        Returns only the current operator's active platform roles and server-computed workspace
        navigation. This bootstrap endpoint does not grant privileged data access or satisfy MFA.
      operationId: getAdminSession
      responses:
        "200":
          description: Authorized operator console projection
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: socal_session
      description: Opaque 256-bit bearer token; host-only, Secure, HttpOnly, SameSite=Lax, Path=/v1. Only its HMAC is persisted.
  parameters:
...
    DeviceId:
      name: X-Device-Id
      in: header
      required: true
      description: Opaque client-generated installation identifier used only through a keyed hash for abuse controls.
      schema: { type: string, minLength: 16, maxLength: 128, pattern: "^[A-Za-z0-9._:-]+$" }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    Unauthorized:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Forbidden:
      description: Permission denied
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    ServiceUnavailable:
      description: Required dependency or feature is unavailable
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  schemas:
...
    ProblemDetails:
      type: object
      required: [type, title, status, detail, instance, requestId]
      examples:
        - type: https://api.socal.local/problems/validation
          title: Bad Request
          status: 400
          detail: Request validation failed
          instance: /v1/listings
          requestId: 00000000-0000-4000-8000-000000000001
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string }
        requestId: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
    OtpRequest:
      type: object
      required: [channel, destination, purpose]
      additionalProperties: false
      properties:
        channel: { type: string, enum: [SMS, EMAIL] }
        destination: { type: string, maxLength: 320 }
        purpose: { type: string, enum: [SIGN_IN, VERIFY_CONTACT, SENSITIVE_ACTION] }
        locale: { type: string, enum: [zh-Hans, en-US], default: zh-Hans }
    OtpVerifyRequest:
      type: object
      required: [challengeId, code]
      additionalProperties: false
      properties:
        challengeId: { type: string, format: uuid }
        code: { type: string, pattern: '^\\d{6}$' }
    OtpAcceptedResponse:
      type: object
      required: [accepted, requestId, challengeId, expiresAt]
      additionalProperties: false
      properties:
        accepted: { type: boolean, const: true }
        requestId: { type: string }
        challengeId: { type: string, format: uuid }
        expiresAt: { type: string, format: date-time }
    SessionResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Session"
    Session:
      type: object
      required: [user, expiresAt, permissions, platformRoles]
      additionalProperties: false
      properties:
        user:
          $ref: "#/components/schemas/UserSummary"
        expiresAt: { type: string, format: date-time }
        permissions:
          type: array
          uniqueItems: true
          items: { type: string }
        platformRoles:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/PlatformRole"
        organizations:
          type: array
          items:
            $ref: "#/components/schemas/OrganizationSummary"
    PlatformRole:
      type: string
      enum:
        [
          SUPPORT,
          MODERATOR,
          SENIOR_MODERATOR,
          AD_OPS,
          FINANCE,
          TAXONOMY_ADMIN,
          PLATFORM_ADMIN,
          READ_ONLY_AUDITOR,
        ]
    AdminNavigationItem:
      type: object
      required: [key, href]
      additionalProperties: false
      properties:
        key:
          type: string
          enum: [moderation, people, taxonomy, commerce, ads, audit, system]
        href:
          type: string
          enum:
            [
              /admin/moderation/listings,
              /admin/users,
              /admin/taxonomy/categories,
              /admin/commerce/orders,
              /admin/ads/campaigns,
              /admin/audit,
              /admin/system/health,
            ]
    AdminSecurityState:
      type: object
      required: [mfaRequired, privilegedActionsAllowed]
      additionalProperties: false
      properties:
        mfaRequired: { type: boolean, const: true }
        privilegedActionsAllowed:
          type: boolean
          description: Remains false until AUTH-005 establishes and verifies an MFA-bound Admin session.
    AdminSession:
      type: object
      required: [operator, roles, navigation, security]
      additionalProperties: false
      properties:
        operator:
          $ref: "#/components/schemas/UserSummary"
        roles:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: "#/components/schemas/PlatformRole"
        navigation:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/AdminNavigationItem"
        security:
          $ref: "#/components/schemas/AdminSecurityState"
    AdminSessionResponse:
      type: object
      required: [data]
      additionalProperties: false
      properties:
        data:
          $ref: "#/components/schemas/AdminSession"
...
    UserSummary:
      type: object
      required: [id, displayName, locale, status]
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string }
        avatarUrl: { type: [string, "null"], format: uri }
        locale: { type: string, enum: [zh-Hans, en-US] }
        status: { type: string, enum: [ACTIVE, LIMITED, SUSPENDED] }
        verificationBadges:
          type: array
          items: { type: string }
    OrganizationSummary:
      type: object
      required: [id, type, displayName, slug, role]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        type:
          $ref: "#/components/schemas/OrganizationType"
        displayName: { type: string }
        slug: { type: string }
        role:
          $ref: "#/components/schemas/MembershipRole"
    OrganizationType:
      type: string
      enum: [MERCHANT, SERVICE_PROVIDER, SUPPLIER, MEDIA, INTERNAL]
    MembershipRole:
      type: string
      enum: [OWNER, ADMIN, EDITOR, BILLING, ANALYST]
...